new convex hull added

+++ src/com/kevmayo/chalkie/android/ButtonDO.java
@@ -19,7 +19,7 @@ public class ButtonDO extends DisplayObject {
 		super("button");
 		_over = over;
 		_out = out;
-		_rect = new Rect(0, 0, _over.size.x, _over.size.y);
+		_rect = new Rect(0, 0, (int) _over.size.x, (int)_over.size.y);
 	}
 
 	@Override
@@ -28,11 +28,11 @@ public class ButtonDO extends DisplayObject {
 		super.draw(g);
 
 		if (_isOver)
-			g.drawImage(_over.image, _rect.left, _rect.top, _over.pos.x,
-					_over.pos.y, _over.size.x, _over.size.y);
+			g.drawImage(_over.image, _rect.left, _rect.top, (int)_over.pos.x,
+					(int)_over.pos.y, (int)_over.size.x, (int)_over.size.y);
 		else
-			g.drawImage(_out.image, _rect.left, _rect.top, _out.pos.x,
-					_out.pos.y, _out.size.x, _out.size.y);
+			g.drawImage(_out.image, _rect.left, _rect.top, (int)_out.pos.x,
+					(int)_out.pos.y, (int)_out.size.x, (int)_out.size.y);
 	}
 
 }
