+++ src/com/kevmayo/chalkie/android/TextureDO.java
@@ -24,7 +24,7 @@ public class TextureDO implements IDisplayObject {
 	
 	public TextureDO(TextureInfo texture){
 		_textureInfo = texture;
-		_rect = new Rect(0, 0, texture.size.x, texture.size.y);
+		_rect = new Rect(0, 0, (int)texture.size.x, (int)texture.size.y);
 	}
 	
 	@Override
@@ -66,7 +66,7 @@ public class TextureDO implements IDisplayObject {
 
 	@Override
 	public void draw(Graphics g) {
-		g.drawImage(_textureInfo.image,_rect.left, _rect.top, _textureInfo.pos.x, _textureInfo.pos.y, _textureInfo.size.x, _textureInfo.size.y);
+		g.drawImage(_textureInfo.image,_rect.left, _rect.top,(int) _textureInfo.pos.x, (int)_textureInfo.pos.y, (int)_textureInfo.size.x, (int)_textureInfo.size.y);
 	}
 
 	@Override
